+++ src/service/cookies.js
@@ -37,7 +37,7 @@ angularServiceInject('$cookies', function($browser) {
   //at the end of each eval, push cookies
   //TODO: this should happen before the "delayed" watches fire, because if some cookies are not
   //      strings or browser refuses to store some cookies, we update the model in the push fn.
-  this.$observe(push);
+  this.$watch(push);
 
   return cookies;